requirements.txt
poetry~=1.8.5
GitPython~=3.1.43
toml~=0.10.2
libcst~=1.5.1
coverage~=7.6.9
radon~=6.0.1
pyflakes~=3.2.0
openai~=1.58.1
PyGithub~=2.5.0
pydantic-settings~=2.7.0